#335 feat: add 24 hour txn volume to home page

+++ apps/explorer/lib/explorer/chain/transaction/history/transaction_stats.ex
@@ -19,7 +19,7 @@ defmodule Explorer.Chain.Transaction.History.TransactionStats do
     field(:number_of_transactions, :integer)
     field(:gas_used, :decimal)
     field(:total_fee, :decimal)
-    field(:today_start_block, :decimal)
+    field(:today_start_block, :integer)
   end
 
   @typedoc """
